Add missing sequence name configuration for the period_id sequence in the upgradestep.
When updating a gever deployment and adding the `periods` table with the [4601 Upgradestep](https://github.com/4teamwork/opengever.core/blob/master/opengever/meeting/upgrades/to4618.py#L12), the commission creation fails with the following error.

```
ProgrammingError: (psycopg2.ProgrammingError) relation "period_id_seq" does not exist
```

This is because the sequence name is only defined in the Period model but not in the upgradestep model copy.