Add related documents method to document overview
... also improve overview's `get_css_class` method.

When adding / editing a document (let's say A), related documents
(sat B and C) can be selected. These are listed under "related
documents" in the document overview (A lists B and C)..
Those related objects should list the documents which are related to
them as well (B and C should display A as well).

See also: #1498