Fix an issue where creating a dossier from template ended in an exception.
While creating a dossier from a template, the user fills out fields on different wizard steps. Each step requires a new request. In the first step, the user chooses which dossiertemplate he wants to add, in the second step he fills out the dossier-fields prefilled from the choosen dossiertemplate and then he can save the dossier. We use an object-storage to retrive the dossietemplate in the next request. In some cases, this object looses its acquisitionwrapper which ends in an exception.

To fix this issue, we changed the implementation to use a path-storage and we retrieve the object in each request through traversing.